Added UX for adding feeds once they have been identified
- When adding feeds, site title and feed url are automatically populated
- Radio buttons are created to select between multiple feeds
- Values:
Control through selecting feeds oneself (the reader does not
automatically grab the first feed or force to either rss or h-feed).
The radio buttons prompt users to check/make a decision before
proceeding.

Ease of use because defaults are selected.

Basically a balance of two values that are at odds with one another.