Like button
Added a like button. Currently it saves the ‘like’ post as a draft
(since I want to be able to test a bit more).

When you click like, it automatically creates a post with content
“Liked <link>”
- After creating the post, post format is set to “link”
- If post kinds plugin is installed, post kind is set to “like”.

Not done yet, but important to do, is adding the ability to unlike a
post. This means ‘likes’ have to be recorded in the database.  	- This
is important for usability.  Basically, the user should not have to go
to the post editor, but should be able to work entirely from the feed
reader interface. Adding this feature supports ease of use

Additionally, The most recent 100 log items from jjreader_log are now
displayed on the settings page. This is just for convenience so I don't
need to look at the database manually.

deleted the file jjreader_responses.php and just put the response
function in the main jjreader.php file. (Maybe I’ll chop up the files
at a later date for clearer organization)