Cleaned up rss feed finding | Re-organized rss/h-feed aggregator
Added an isRSS() function which returns true if a feed is of a known
rss type (or atom etc.).  This makes it simpler to ID rss feeds

Edited the aggregator so it has a place to process h-feeds.  Currently
this section just logs “[feed url] is an h-feed”, but this is where the
h-feed scraping code will go