Still working on h-feeds
Improved detection of redundant titles (e.g. for posts that are not
intended to have a title, so  ‘name’ is the same as ‘content’ or
‘summary’.
- Notably, for h-feeds I check the title against both content[html] and
content[value].

So far this seems successful at preventing redundant titles from being
loaded.

Fixed a bug where, for h-feeds, I was sending an empty variable ‘time’
instead of the proper variable ‘published’.  This meant that h-feed
posts were being read with no published date.

Found a bug where h-feed publication datetimes with timezone info are
not stored properly. When adding posts they are converted using date(),
which cannot parse the timezone info (at least I think this is the
cause). Thinking of solutions now.
Option 1) Store timezone offset in a separate variable, and convert
(and sort) all datetimes to UTC.  Process the timezone offset on the
front-end.
Option 2) discard timezone information and convert all datetimes to the
user’s default timezone.  ( I think this is worse, because the timezone
info can be important — i.e. noting that someone is travelling)

Option 3) Look into php datetime — maybe it’s possible to store each
date as a datetime object including the timezone.