Improving post display
Database tables
- Changed collation of text columns to utf8mb4_unicode_ci — this makes
it possible to store emoji and other special characters

Summary display
- Added more tags to be kept when running strip_tags on the summary.
- Was pleased to learn that wordpress seems to automatically closes any
tags that are left open by strip_tags. Keep an eye on this.  If there
are any errors I might be able to fix this with php tidy

Full content display
- Increased z-index of the full content pop-up to 999. This should make
it appear in front of any reasonably formatted page. It’s possible some
themes might still break this but I’m not sure how to get around that.