Improved click to close full screen view
Full screen view should close when user clicks on the close button or
the background (outside the content).

Previously I prevented clicks within the content div from closing this
view by using stopPropagation(). But this prevented clicks inside the
div from triggering as well, such as clicking the ‘like’ or ‘reply’
buttons.

So changed to a new message based on
https://stackoverflow.com/questions/16159229/lightbox-click-outside-imag
e-to-close