Improved click to close full content view
Fixed a bug where clicks inside the full-content box were disabled
(including links).  Bug was because I was trying to allow clicks on the
background the close the full-content box, whereas clicks with the box
itself should not close it.  Previous approach was to “return false;”
when the user clicks inside the content box. New approach compared
target and currentTarget, and only proceeds if they match.

Solution courtesy of:
https://stackoverflow.com/a/41621023