In progress - improving h-feed parsing
- Store full array of photos instead of just the first one (Even though
yarns only displays the first photo in the summary, storing the full
array adds flexibility.

- Change h-feed fetching flow:
1) Fetch permalinks from h-feed
2) Fetch individual posts from permalinks
    - This is an improvement because many sites have incomplete mf2
markup in the feed, but complete markup on individual posts

- Changed ‘photo’ in the _posts Table to mediumtext (to store long
arrays) - Maybe unnecessary.

- Cleaned up the code a bit (removed old cron timers and some
deprecated code)

- Added a check that removes ‘name’ property for feed items when name
== content[0][‘value’].  (Previously this was just done when adding
items, but it makes sense to do it earlier in the parse loop. May want
to clean up the yarns_reader_add_feeditem if this check is redundant
there)

Still to do: Improve parsing of h-cards for in-reply-to, author, and
location fields.