Parse author h-cards if available
For h-feed items, check if ‘author’ contains an h-card. If so, fetch
author name, photo and url.  if not, just fetch ‘value’

For the moment, author name is not used (the user-determined site title
is displayed instead) and author url is omitted in favour of site url.
Maybe I’ll change this later.  Preserving those properties for now just
in case.

I also removed a few comments (mostly old calls to yarns_reader_log I
was using for debugging before)