improved logic for displaying feed items
Further Improved logic for comparing content and summaries (previously
there were many false negatives when trying to find duplication).
(There is some redundancy because parts of this check occur when adding
items to the database, and parts when displaying posts. Need to figure
out a more coherent plan here)

Created function for cleaning and then comparing two strings. (This
replaces a title vs. content specific method with a more generic
version, which I now use to compare summary and content as well)

Added a couple css tweaks to improve the display of in-reply-to links