Parse locations as h-cards if available
If locations are presented as h-cards, get name, url, latitude and
longitude if available.

In the feed reader display, locations are displayed as links if
possible, or plain text otherwise.  latitude and longitude are not
currently used, but I’m capturing them for now anyway so I can use them
in the future